api/requirements.txt
minio
python-dotenv
PyPDF2